Struct jsonrpc_core::IoHandler

pub struct IoHandler<M: Metadata = ()>(_);

Simplified IoHandler with no Metadata associated with each request.

Implementations

impl IoHandler

pub fn new() -> Self

Creates new IoHandler without any metadata.

pub fn with_compatibility(compatibility: Compatibility) -> Self

Creates new IoHandler without any metadata compatible with specified protocol version.

impl<M: Metadata + Default> IoHandler<M>

pub fn handle_request(
&self,
request: &str
) -> FutureResult<FutureResponse, FutureOutput>

Handle given string request asynchronously.

pub fn handle_rpc_request(
&self,
request: Request
) -> FutureRpcResult<FutureResponse, FutureOutput>

Handle deserialized RPC request asynchronously.

pub fn handle_call(
&self,
call: Call
) -> Either<FutureOutput, Either<FutureOutput, Ready<Option<Output>>>>

Handle single Call asynchronously.

pub fn handle_request_sync(&self, request: &str) -> Option<String>

Handle given request synchronously - will block until response is available. If you have any asynchronous methods in your RPC it is much wiser to use handle_request instead and deal with asynchronous requests in a non-blocking fashion.