Struct jsonrpc_core::IoHandler

pub struct IoHandler<M: Metadata = ()>(_);

Simplified IoHandler with no Metadata associated with each request.

Methods

impl IoHandler

fn new() -> Self

Creates new IoHandler without any metadata.

fn with_compatibility(compatibility: Compatibility) -> Self

Creates new IoHandler without any metadata compatible with specified protocol version.

impl<M: Metadata> IoHandler<M>

fn handle_request(&self, request: &str) -> BoxFuture<Option<String>, ()>

Handle given string request asynchronously.

fn handle_rpc_request(&self, request: Request) -> FutureResponse

Handle deserialized RPC request asynchronously.

fn handle_call(&self, call: Call) -> BoxFuture<Option<Output>, ()>

Handle single Call asynchronously.

fn handle_request_sync(&self, request: &str) -> Option<String>

Handle given request synchronously - will block until response is available. If you have any asynchronous methods in your RPC it is much wiser to use handle_request instead and deal with asynchronous requests in a non-blocking fashion.