Module jsonrpc_core::types

JSON-RPC types

Re-exports

pub use self::error::Error;

pub use self::error::ErrorCode;

pub use self::id::Id;

pub use self::params::Params;

pub use self::request::Call;

pub use self::request::MethodCall;

pub use self::request::Notification;

pub use self::request::Request;

pub use self::response::Failure;

pub use self::response::Output;

pub use self::response::Response;

pub use self::response::Success;

pub use self::version::Version;

Modules

error

jsonrpc errors

id

jsonrpc id field

params

jsonrpc params field

request

jsonrpc request

response

jsonrpc response

version

jsonrpc version field

Enums

Value

Represents any valid JSON value.

Functions

to_string

Serialize the given data structure as a String of JSON.

to_value

Convert a T into serde_json::Value which is an enum that can represent any valid JSON data.